requirements/base.txt
Faker==1.0.7
Django==1.11.10
django-extensions==1.9.9
arrow==0.12.1
django_validator==0.2.7
jsonfield==2.0.2
djangorestframework==3.8.2
django-model-utils==3.1.2

social-auth-app-django==2.1.0
social-auth-core==1.7.0
wechatpy==1.7.3
cryptography==2.2.2
django-oauth-toolkit==1.1

ipython<6
pylint==1.8.2
pylint-django==0.9.0
pylint-plugin-utils==0.2.6
django-nose==1.4.5
mock==2.0.0
factory-boy==2.10.0
ddt==1.1.1
pep8==1.7.1
mysqlclient==1.3.12
qiniu==7.2.2
requests==2.13.0
certifi==2015.9.6.2
django-simple-history==1.9.0

qcloudsms-py==0.1.1
yunpian-python-sdk==1.0.0
Pillow==5.2.0
path.py==11.0.1
PyQRCode==1.2.1
python-memcached==1.59
gunicorn==19.9.0
beautifulsoup4==4.7.1

-e git+https://github.com/duoduo369/cos-python-sdk-v5.git@732c4b3357933421b07#egg=cos-python-sdk-v5
-e git+https://github.com/duoduo369/django-mako.git@1177157317522edcc56d55a530dd3d7547dbb823#egg=django-mako
